requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
drf-yasg==1.17.1
